# image-reactor/doc/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image Processor API v1",
    "summary": "Image metadata processor with FastAPI and asyncpg",
    "description": "Image metadata processor API.\n\n## Features\n- Store and retrieve image metadata\n- Track failed download attempts\n- Health check endpoints\n",
    "contact": {
      "name": "Jan Trojak",
      "email": "jan.trojak@recombee.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/images": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Create image",
        "description": "Create a new image record in the database. The stored_at timestamp will be automatically set to NOW().",
        "operationId": "create_image_images_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/by-path": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get image by path",
        "description": "Retrieve an image by its store_collection and filepath",
        "operationId": "get_image_by_path_images_by_path_get",
        "parameters": [
          {
            "name": "store_collection",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Storage collection name",
              "title": "Store Collection"
            },
            "description": "Storage collection name"
          },
          {
            "name": "filepath",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "File path within the collection",
              "title": "Filepath"
            },
            "description": "File path within the collection"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/{image_id}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get image by ID",
        "description": "Retrieve an image by its database ID",
        "operationId": "get_image_by_id_images__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Images"
        ],
        "summary": "Delete image by ID",
        "description": "Delete an image by its ID. This will also cascade delete all associated download attempts.",
        "operationId": "delete_image_images__image_id__delete",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Image Images  Image Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Image": {
        "properties": {
          "database_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Id",
            "description": "External database identifier"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id",
            "description": "External item identifier"
          },
          "property_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name",
            "description": "Name of the property"
          },
          "image_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Number",
            "description": "Image number within the property"
          },
          "link": {
            "type": "string",
            "title": "Link",
            "description": "Image URL"
          },
          "store_collection": {
            "type": "string",
            "maxLength": 256,
            "title": "Store Collection",
            "description": "Storage collection name"
          },
          "filepath": {
            "type": "string",
            "title": "Filepath",
            "description": "Local file path"
          },
          "hashsum": {
            "type": "string",
            "title": "Hashsum",
            "description": "File hash (hex string)"
          },
          "extension": {
            "type": "string",
            "maxLength": 32,
            "title": "Extension",
            "description": "File extension"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "description": "Image width in pixels"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "description": "Image height in pixels"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "stored_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored At",
            "description": "Storage timestamp"
          }
        },
        "type": "object",
        "required": [
          "link",
          "store_collection",
          "filepath",
          "hashsum",
          "extension",
          "width",
          "height",
          "id"
        ],
        "title": "Image",
        "description": "Complete image model with database fields."
      },
      "ImageCreate": {
        "properties": {
          "database_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Id",
            "description": "External database identifier"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id",
            "description": "External item identifier"
          },
          "property_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name",
            "description": "Name of the property"
          },
          "image_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Number",
            "description": "Image number within the property"
          },
          "link": {
            "type": "string",
            "title": "Link",
            "description": "Image URL"
          },
          "store_collection": {
            "type": "string",
            "maxLength": 256,
            "title": "Store Collection",
            "description": "Storage collection name"
          },
          "filepath": {
            "type": "string",
            "title": "Filepath",
            "description": "Local file path"
          },
          "hashsum": {
            "type": "string",
            "title": "Hashsum",
            "description": "File hash (hex string)"
          },
          "extension": {
            "type": "string",
            "maxLength": 32,
            "title": "Extension",
            "description": "File extension"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "description": "Image width in pixels"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "description": "Image height in pixels"
          }
        },
        "type": "object",
        "required": [
          "link",
          "store_collection",
          "filepath",
          "hashsum",
          "extension",
          "width",
          "height"
        ],
        "title": "ImageCreate",
        "description": "Model for creating a new image record."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}